+++ src/host/enclave.h
@@ -10,17 +10,23 @@
 
 #include <dlfcn.h>
 #include <filesystem>
-#ifdef VIRTUAL_ENCLAVE
-#  include "enclave/ccf_v.h"
-#else
+
+#ifdef CCHOST_SUPPORTS_SGX
 #  include <ccf_u.h>
 #  include <openenclave/bits/result.h>
 #  include <openenclave/host.h>
 #  include <openenclave/trace.h>
 #endif
 
+#ifdef CCHOST_SUPPORTS_VIRTUAL
+// Include order matters. virtual_enclave.h uses the OE definitions if
+// available, else creates its own stubs
+#  include "enclave/virtual_enclave.h"
+#endif
+
 extern "C"
 {
+#ifdef CCHOST_SUPPORTS_SGX
   void nop_oe_logger(
     void* context,
     bool is_enclave,
@@ -30,14 +36,40 @@ extern "C"
     uint64_t host_thread_id,
     const char* message)
   {}
+#endif
 }
-
-// Marker to create virtual enclaves, should be distinct from any valid
-// OE_ENCLAVE_FLAG combinations
-constexpr static uint32_t ENCLAVE_FLAG_VIRTUAL = -1;
 
 namespace host
 {
+  void expect_enclave_file_suffix(
+    const std::string& file,
+    char const* expected_suffix,
+    host::EnclaveType type)
+  {
+    if (!nonstd::ends_with(file, expected_suffix))
+    {
+      // Remove possible suffixes to try and get root of filename, to build
+      // suggested filename
+      auto basename = file;
+      for (const char* suffix :
+           {".signed", ".debuggable", ".so", ".enclave", ".virtual"})
+      {
+        if (nonstd::ends_with(basename, suffix))
+        {
+          basename = basename.substr(0, basename.size() - strlen(suffix));
+        }
+      }
+      const auto suggested = fmt::format("{}{}", basename, expected_suffix);
+      throw std::logic_error(fmt::format(
+        "Given enclave file '{}' does not have suffix expected for enclave "
+        "type "
+        "{}. Did you mean '{}'?",
+        file,
+        nlohmann::json(type).dump(),
+        suggested));
+    }
+  }
+
   /**
    * Wraps an oe_enclave and associated ECalls. New ECalls should be added as
    * methods which construct an appropriate EGeneric-derived param type and pass
@@ -46,22 +78,22 @@ namespace host
   class Enclave
   {
   private:
-    bool is_virtual_enclave;
-
-    oe_enclave_t* e;
+#ifdef CCHOST_SUPPORTS_SGX
+    oe_enclave_t* sgx_handle = nullptr;
+#endif
+#ifdef CCHOST_SUPPORTS_VIRTUAL
+    void* virtual_handle = nullptr;
+#endif
 
   public:
     /**
      * Create an uninitialized enclave hosting the given library.
      *
      * @param path Path to signed enclave library file
-     * @param flags Flags passed to oe_create_enclave. eg OE_ENCLAVE_FLAG_DEBUG,
-     * OE_ENCLAVE_FLAG_SIMULATE. Alternatively, ENCLAVE_FLAG_VIRTUAL will not
-     * use OE at all, instead loading a shared library directly
+     * @param type Type of enclave to load, influencing what flags should be
+     * passed to OE, or whether to dlload a virtual enclave
      */
-    Enclave(const std::string& path, uint32_t flags) :
-      is_virtual_enclave(false),
-      e(nullptr)
+    Enclave(const std::string& path, EnclaveType type)
     {
       if (!std::filesystem::exists(path))
       {
@@ -69,26 +101,63 @@ namespace host
           fmt::format("No enclave file found at {}", path));
       }
 
-      if (flags == ENCLAVE_FLAG_VIRTUAL)
+      switch (type)
       {
-#ifdef VIRTUAL_ENCLAVE
-        load_virtual_enclave(path.c_str());
-#endif
-        is_virtual_enclave = true;
-      }
-      else
-      {
-#ifndef VERBOSE_LOGGING
-        oe_log_set_callback(nullptr, nop_oe_logger);
-#endif
-
-        auto err = oe_create_ccf_enclave(
-          path.c_str(), OE_ENCLAVE_TYPE_SGX, flags, nullptr, 0, &e);
-
-        if (err != OE_OK)
+        case host::EnclaveType::SGX_RELEASE:
+        case host::EnclaveType::SGX_DEBUG:
         {
+#ifdef CCHOST_SUPPORTS_SGX
+          uint32_t oe_flags = 0;
+          if (type == host::EnclaveType::SGX_DEBUG)
+          {
+            expect_enclave_file_suffix(path, ".enclave.so.debuggable", type);
+            oe_flags |= OE_ENCLAVE_FLAG_DEBUG;
+          }
+          else
+          {
+            expect_enclave_file_suffix(path, ".enclave.so.signed", type);
+          }
+
+#  ifndef VERBOSE_LOGGING
+          oe_log_set_callback(nullptr, nop_oe_logger);
+#  endif
+
+          auto err = oe_create_ccf_enclave(
+            path.c_str(),
+            OE_ENCLAVE_TYPE_SGX,
+            oe_flags,
+            nullptr,
+            0,
+            &sgx_handle);
+
+          if (err != OE_OK)
+          {
+            throw std::logic_error(
+              fmt::format("Could not create enclave: {}", oe_result_str(err)));
+          }
+#else
           throw std::logic_error(
-            fmt::format("Could not create enclave: {}", oe_result_str(err)));
+            "SGX enclaves are not supported in current build");
+#endif // CCHOST_SUPPORTS_SGX
+          break;
+        }
+
+        case host::EnclaveType::VIRTUAL:
+        {
+#ifdef CCHOST_SUPPORTS_VIRTUAL
+          expect_enclave_file_suffix(path, ".virtual.so", type);
+          virtual_handle = load_virtual_enclave(path.c_str());
+#else
+          throw std::logic_error(
+            "Virtual enclaves not supported in current build");
+#endif // CCHOST_SUPPORTS_VIRTUAL
+          break;
+        }
+
+        default:
+        {
+          throw std::logic_error(fmt::format(
+            "Unsupported enclave type: {}", nlohmann::json(type).dump()));
         }
       }
     }
@@ -102,7 +171,7 @@ namespace host
       size_t num_worker_thread,
       void* time_location)
     {
-      CreateNodeStatus status;
+      CreateNodeStatus status = CreateNodeStatus::InternalError;
       constexpr size_t enclave_version_size = 256;
       std::vector<uint8_t> enclave_version_buf(enclave_version_size);
 
@@ -112,24 +181,29 @@ namespace host
 
       auto config = nlohmann::json(ccf_config).dump();
 
-      auto err = enclave_create_node(
-        e,
-        &status,
-        (void*)&enclave_config,
-        config.data(),
-        config.size(),
-        node_cert.data(),
-        node_cert.size(),
-        &node_cert_len,
-        service_cert.data(),
-        service_cert.size(),
-        &service_cert_len,
-        enclave_version_buf.data(),
-        enclave_version_buf.size(),
-        &enclave_version_len,
-        start_type,
-        num_worker_thread,
-        time_location);
+#define CREATE_NODE_ARGS \
+  &status, (void*)&enclave_config, config.data(), config.size(), \
+    node_cert.data(), node_cert.size(), &node_cert_len, service_cert.data(), \
+    service_cert.size(), &service_cert_len, enclave_version_buf.data(), \
+    enclave_version_buf.size(), &enclave_version_len, start_type, \
+    num_worker_thread, time_location
+
+      oe_result_t err = OE_FAILURE;
+
+// Assume that constructor correctly set the appropriate field, and call
+// appropriate function
+#ifdef CCHOST_SUPPORTS_VIRTUAL
+      if (virtual_handle != nullptr)
+      {
+        err = virtual_create_node(virtual_handle, CREATE_NODE_ARGS);
+      }
+#endif
+#ifdef CCHOST_SUPPORTS_SGX
+      if (sgx_handle != nullptr)
+      {
+        err = enclave_create_node(sgx_handle, CREATE_NODE_ARGS);
+      }
+#endif
 
       if (err != OE_OK || status != CreateNodeStatus::OK)
       {
@@ -163,7 +237,20 @@ namespace host
     bool run()
     {
       bool ret;
-      auto err = enclave_run(e, &ret);
+      oe_result_t err = OE_FAILURE;
+
+#ifdef CCHOST_SUPPORTS_VIRTUAL
+      if (virtual_handle != nullptr)
+      {
+        err = virtual_run(virtual_handle, &ret);
+      }
+#endif
+#ifdef CCHOST_SUPPORTS_SGX
+      if (sgx_handle != nullptr)
+      {
+        err = enclave_run(sgx_handle, &ret);
+      }
+#endif
 
       if (err != OE_OK)
       {