+++ src/node/hooks.h
@@ -4,9 +4,9 @@
 
 #include "ds/logger.h"
 #include "enclave/reconfiguration_type.h"
-#include "node/config.h"
-#include "node/signatures.h"
-#include "node_info_network.h"
+#include "service/tables/config.h"
+#include "service/tables/node_info_network.h"
+#include "service/tables/signatures.h"
 
 namespace ccf
 {