+++ src/node/test/history.cpp
@@ -8,8 +8,9 @@
 #include "kv/test/null_encryptor.h"
 #include "kv/test/stub_consensus.h"
 #include "node/entities.h"
-#include "node/nodes.h"
-#include "node/signatures.h"
+#include "service/blit.h"
+#include "service/tables/nodes.h"
+#include "service/tables/signatures.h"
 
 #define DOCTEST_CONFIG_IMPLEMENT
 #include <doctest/doctest.h>