+++ src/service/genesis_gen.h
@@ -2,14 +2,13 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 #include "ccf/tx.h"
-#include "code_id.h"
 #include "crypto/hash.h"
 #include "crypto/verifier.h"
-#include "entities.h"
-#include "ledger_secrets.h"
-#include "members.h"
 #include "network_tables.h"
-#include "nodes.h"
+#include "node/ledger_secrets.h"
+#include "service/tables/code_id.h"
+#include "service/tables/members.h"
+#include "service/tables/nodes.h"
 
 #include <algorithm>
 #include <fstream>